pallet-grandpa-asp: move tests and vote trait to their own files
